fix: lint
Jira: CHAOS-577
